refactor: criteria 레이어 변경 (domain -> application)
